feat: add floating theme switcher on home page
